+++ packages/mirrorinput/src/constants.ts
@@ -1,4 +1,4 @@
-import { createDynamicAttibuteValue } from '$utils/attributes';
+import { generateDynamicAttibuteValue, generateSelectors } from '$utils/attributes';
 
 const ATTRIBUTES_PREFIX = 'fs-mirrorinput';
 
@@ -17,7 +17,9 @@ export const ATTRIBUTES = {
       /**
        * Defines the element as the target to mirror the fired event.
        */
-      target: createDynamicAttibuteValue('target'),
+      target: generateDynamicAttibuteValue('target'),
     },
   },
 } as const;
+
+export const getSelector = generateSelectors(ATTRIBUTES);