+++ packages/countitems/CHANGELOG.md
@@ -1,5 +1,9 @@
 # `countitems` Changelog
 
+## [v1.0.1] 25th September 2021
+
+- Implemented the `generateSelectors` util.
+
 ## [v1.0.0] 25th September 2021
 
 - Created `CHANGELOG.md`.