+++ packages/linkblockedit/CHANGELOG.md
@@ -1,5 +1,9 @@
 # `linkblockedit` Changelog
 
+## [v1.0.2] 25th September 2021
+
+- Implemented the `generateSelectors` util.
+
 ## [v1.0.1] 25th September 2021
 
 - Fixed `currentScript` vs `params` distinction.