+++ packages/mirrorclick/CHANGELOG.md
@@ -1,5 +1,9 @@
 # `mirrorclick` Changelog
 
+## [v1.1.1] 25th September 2021
+
+- Implemented the `generateSelectors` util.
+
 ## [v1.1.0] 25th September 2021
 
 - Created `CHANGELOG.md`.