+++ packages/scrolldisable/CHANGELOG.md
@@ -1,5 +1,9 @@
 # `scrolldisable` Changelog
 
+## [v1.1.6] 25th September 2021
+
+- Implemented the `generateSelectors` util.
+
 ## [v1.1.5] 25th September 2021
 
 - Fixed `currentScript` vs `params` distinction.