+++ packages/untransform/CHANGELOG.md
@@ -1,5 +1,9 @@
 # `untransform` Changelog
 
+## [v1.0.6] 25th September 2021
+
+- Implemented the `generateSelectors` util.
+
 ## [v1.0.5] 25th September 2021
 
 - Created `CHANGELOG.md`.